DataPointDwr

+++ Core/src/com/serotonin/m2m2/web/dwr/DataPointDwr.java
@@ -26,6 +26,7 @@ import com.serotonin.m2m2.LicenseViolatedException;
 import com.serotonin.m2m2.db.dao.DataPointDao;
 import com.serotonin.m2m2.db.dao.EventDetectorDao;
 import com.serotonin.m2m2.db.dao.ResultsWithTotal;
+import com.serotonin.m2m2.db.dao.SystemSettingsDao;
 import com.serotonin.m2m2.i18n.ProcessResult;
 import com.serotonin.m2m2.i18n.TranslatableMessage;
 import com.serotonin.m2m2.module.EventDetectorDefinition;
@@ -63,7 +64,7 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
         LOG = LogFactory.getLog(DataPointDwr.class);
     }
 
-    @DwrPermission(user = true)
+    @DwrPermission(custom = SystemSettingsDao.PERMISSION_DATASOURCE)
     public ProcessResult getPoints() {
         ProcessResult result = new ProcessResult();
 
@@ -140,6 +141,7 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
         User user = Common.getUser();
 
         if (id == Common.NEW_ID) {
+            Permissions.ensureDataSourcePermission(user);
             vo = dao.getNewVo();
             //TODO Need to sort this out another way, this will wreck 
             // when users have mulitple tabs open in a browser
@@ -156,11 +158,10 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
         }
         else {
             vo = dao.getFull(id);
+            if(vo != null)
+                Permissions.ensureDataSourcePermission(user, vo.getDataSourceId());
         }
 
-        //Should check permissions?
-        //Permissions.ensureDataSourcePermission(user, vo.getDataSourceId());
-        user.setEditPoint(vo);
         //TODO Need to deal with point value defaulter
 
         ProcessResult response = new ProcessResult();
@@ -181,8 +182,10 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
         ProcessResult response = new ProcessResult();
         try {
             DataPointVO dp = dao.get(id);
-            if (dp != null)
+            if (dp != null) {
+                Permissions.ensureDataSourcePermission(Common.getHttpUser(), dp.getId());
                 Common.runtimeManager.deleteDataPoint(dp);
+            }
         }
         catch (Exception e) {
             // Handle the exceptions.
@@ -215,6 +218,8 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
         if (vo.getXid() == null) {
             vo.setXid(dao.generateUniqueXid());
         }
+        
+        Permissions.ensureDataSourcePermission(Common.getHttpUser(), vo.getDataSourceId());
         vo.validate(response);
 
         if (!response.getHasMessages()) {
@@ -266,6 +271,7 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
 
         //Store the edit point
         DataPointVO editPoint = (DataPointVO) result.getData().get("vo");
+        Permissions.ensureDataSourcePermission(Common.getHttpUser(), editPoint.getDataSourceId());
         editPoint.setId(COPY_ID);
         for(AbstractPointEventDetectorVO<?> aed : editPoint.getEventDetectors()) {
             aed.setId(Common.NEW_ID);
@@ -357,8 +363,10 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
             int purgeType, int purgePeriod, int defaultCacheSize, boolean overrideIntervalLoggingSamples,
             int intervalLoggingSampleWindowSize) {
 
-        DataPointVO dp = Common.getUser().getEditPoint();
+        User user = Common.getUser();
+        DataPointVO dp = user.getEditPoint();
         if (dp != null) {
+            Permissions.ensureDataSourcePermission(user, dp.getDataSourceId());
             dp.setLoggingType(type);
             dp.setIntervalLoggingPeriod(period);
             dp.setIntervalLoggingPeriodType(periodType);
@@ -387,8 +395,10 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
     @SuppressWarnings("deprecation")
     @DwrPermission(user = true)
     public void storeEditProperties(DataPointVO newDp) {
-        DataPointVO dp = Common.getUser().getEditPoint();
+        User user = Common.getUser();
+        DataPointVO dp = user.getEditPoint();
         if (dp != null) {
+            Permissions.ensureDataSourcePermission(user, dp.getDataSourceId());
             //Do we want the details set here? (The ID Name,XID and Locator are stored via the modules)
             dp.setId(newDp.getId());
             dp.setXid(newDp.getXid());
@@ -522,6 +532,8 @@ public class DataPointDwr extends AbstractDwr<DataPointVO, DataPointDao> {
      */
     @DwrPermission(user = true)
     public ProcessResult getMostRecentValue(int id) {
+        DataPointVO dp = DataPointDao.getInstance().get(id);
+        Permissions.ensureDataPointReadPermission(Common.getHttpUser(), dp);
         ProcessResult result = new ProcessResult();
 
         if (Common.runtimeManager.isDataPointRunning(id)) {
